Implement support for EXT_disjoint_timer_query_webgl2.
Until now, it worked with the core query APIs, except for
glQueryCounterEXT() / glGetQueryObjecti64vEXT() -- there, instead of
passing a WebGLQuery object on WebGL 2, it referenced a non-existent
WebGLTimerQueryEXT object as the query was created with
glGenQueries() instead and put into a different array.

Because code coming from GLES may expect the EXT-suffixed versions still
being available on WebGL 2 builds (as you can use the EXT-suffixed
versions of core ES3 APIs with EXT_disjoint_time_query as well), those
are now aliases to the core APIs.